+++ src/meive.jl
@@ -1,20 +1,17 @@
-module Estimator
-
-export eive
-export EiveResult
+module Eivem
 
 
+export meive
+
+import ..Estimator: EiveResult
 import ..CGA: cga
 import Random: AbstractRNG, MersenneTwister
 
-struct EiveResult
-    betas::Vector{Float64}
-end
 
 """
     eive(;
-    dirtyx::Array{T,1},
-    y::Array{T,1},
+    dirtyx::Array{T, 1},
+    y::Array{T, 2},
     otherx::Union{Nothing,Array{T,2},Array{T,1}},
     popsize::Int = 50,
     numdummies::Int = 10,
@@ -22,15 +19,15 @@ end
 )::EiveResult where {T<:Real}
 
 # Description:
-The method searches for a set of dummy (binary) variables that separates the erroneous independent variable
-into clean part and error part. The clean part is then used in the main regression estimation. 
-Those dummy variables minimize the sum of squares of residuals of the main regression. In other terms
-the methods searches for a set of proxy variables that do not exist in real. Please see the reference for details.  
+This is the multivariate case of eive(). Please see eive() function. 
+In the multivariate case, the y is not vector, but a matrix of multiple 
+or repeated measurements of the response variable. This can be considered 
+as multivariate regression as well as regressions with repeated measurements.  
 
 # Arguments:
 
 - dirtyx: Independent variable measured with some error
-- y: Dependent variable
+- y: nxp matrix of dependent variables where n is the number of observations and p is the number of dependent variables
 - otherx: Matrix of other independent variables
 - popsize: Number of individuals in the population (optional)
 - numdummies: Number of dummy variables to use (optional)
@@ -45,19 +42,23 @@ julia> rng = Random.MersenneTwister(1234)
 julia> n = 30
 julia> deltax = randn(rng, n) * sqrt(3.0)
 julia> cleanx = randn(rng, n) * sqrt(7.0)
-julia> e = randn(rng, n) * sqrt(5.0)
-julia> y = 20.0 .+ 10.0 .* cleanx .+ e
+julia> e1 = randn(rng, n) * sqrt(5.0)
+julia> e2 = randn(rng, n) * sqrt(5.0)
+julia> y1 = 20.0 .+ 10.0 .* cleanx .+ e1
+julia> y2 = 10.0 .+ 15.0 .* cleanx .+ e2
 julia> dirtyx = cleanx + deltax
-julia> eive(dirtyx = dirtyx, y = y, otherx = nothing) 
 
-EiveResult([20.28458307772922, 9.456757289676714])
+julia> # Getting bias-reduced estimates
+julia> meive(dirtyx = dirtyx, y = hcat(y1, y2), otherx = nothing) 
+
+EiveResult([19.65449584842238, 9.21108792897651])
 
 julia> X = hcat(ones(n), dirtyx);
 
 julia> # Biased OLS estimates:
-julia> X \\ y
+julia> X \\ y1
 2-element Vector{Float64}:
- 17.94867860059858
+17.94867860059858
   5.8099584879737876
 ```
 
@@ -68,30 +69,23 @@ regression using compact genetic algorithms." Journal of Statistical Computation
  85.16 (2015): 3216-3235.
 
 """
-function eive(;
-    dirtyx::Array{T,1},
-    y::Array{T,1},
+function meive(;
+    dirtyx::Array{T, 1},
+    y::Array{T, 2},
     otherx::Union{Nothing,Array{T,2},Array{T,1}},
     popsize::Int = 50,
     numdummies::Int = 10,
-    rng::AbstractRNG = MersenneTwister(1234)
-)::EiveResult where {T<:Real}
+    rng::AbstractRNG = MersenneTwister(1234))::EiveResult where {T<:Real}
 
     if isnothing(otherx)
-        return eivewithoutotherx(dirtyx, y, popsize, numdummies, rng)
+        return meivewithoutotherx(dirtyx, y, popsize, numdummies, rng)
     else
-        return eivewithotherx(dirtyx, y, otherx, popsize, numdummies, rng)
+        return meivewithotherx(dirtyx, y, otherx, popsize, numdummies, rng)
     end
 end
 
 
-function eivewithotherx(
-    dirtyx::Array{T,1},
-    y::Array{T,1},
-    otherx::Union{Array{T,2},Array{T,1}},
-    popsize::Int = 50,
-    numdummies::Int = 10,
-    rng::AbstractRNG = MersenneTwister(1234))::EiveResult where {T<:Real}
+function meivewithotherx(dirtyx::Array{T,1},y::Array{T,2}, otherx::Union{Array{T,2},Array{T,1}}, popsize::Int = 50, numdummies::Int = 10, rng::AbstractRNG = MersenneTwister(1234))::EiveResult where {T<:Real}
 
 
     n = length(dirtyx)
@@ -106,9 +100,15 @@ function eivewithotherx(
 
         X = hcat(myones, cleanX, otherx)
 
-        outerbetas = X \ y
-        res = y .- X * outerbetas
-        return sum(res .^ 2.0)
+        _ , np = size(y)
+        totalres = 0.0
+        for i in 1:np
+            currenty = y[:, i]
+            outerbetas = X \ currenty
+            res = currenty .- X * outerbetas
+            totalres += sum(res .^ 2.0) 
+        end 
+        return totalres
     end
 
     finalbits = cga(chsize = chsize, costfunction = costfn, popsize = popsize, rng = rng)
@@ -119,20 +119,13 @@ function eivewithotherx(
 
     X = hcat(myones, cleanX, otherx)
 
-    outerbetas = X \ y
+    outerbetas = X \ y[:,1]
 
     return EiveResult(outerbetas)
 end
 
 
-function eivewithoutotherx(
-    dirtyx::Array{T,1},
-    y::Array{T,1},
-    popsize::Int = 50,
-    numdummies::Int = 10,
-    rng::AbstractRNG = MersenneTwister(1234)
-)::EiveResult where {T<:Real}
-
+function meivewithoutotherx(dirtyx::Array{T,1},y::Array{T,2},popsize::Int = 50,numdummies::Int = 10,rng::AbstractRNG = MersenneTwister(1234))::EiveResult where {T<:Real}
 
     n = length(dirtyx)
     myones = ones(Float64, n)
@@ -146,9 +139,15 @@ function eivewithoutotherx(
 
         X = hcat(myones, cleanX)
 
-        outerbetas = X \ y
-        res = y .- X * outerbetas
-        return sum(res .^ 2.0)
+        _ , np = size(y)
+        totalres = 0.0
+        for i in 1:np
+            currenty = y[:, i]
+            outerbetas = X \ currenty
+            res = currenty .- X * outerbetas
+            totalres += sum(res .^ 2.0)
+        end 
+        return totalres
     end
 
     finalbits = cga(chsize = chsize, costfunction = costfn, popsize = popsize, rng = rng)
@@ -159,10 +158,10 @@ function eivewithoutotherx(
 
     X = hcat(myones, cleanX)
 
-    outerbetas = X \ y
+    outerbetas = X \ y[:, 1]
 
     return EiveResult(outerbetas)
 end
 
 
-end # end of module
+end # end of module eivem