+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.1.1 (Upcoming release)
 
+- Add multivariate case (more than one response variables for additional info)
 - Add basic documentation 
 - Add continous integration to repository
 - Add custom RNG for stochastic methods